% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{alcuse}
\alias{alcuse}
\title{Alcohol use among youth}
\format{A data frame with 246 rows and 9 variables:
\describe{
  \item{ID}{ID of participant}
  \item{AGE}{AGE of participant}
  \item{COA}{Are they a child of an alcoholic? 1=yes}
  \item{MALE}{Gender (1=male)}
  \item{AGE_14}{Number of years since age 14}
  \item{ALCUSE}{Average number of alcoholic drinks a week }
  \item{PEER}{Average number of alcoholic drinks a week among their peers}
  \item{CPEER}{Median-centered amount their peers drink}
  \item{CCOA}{Child of alcoholic, centered}
}}
\usage{
alcuse
}
\description{
A dataset containing alcohol use from 82 youth over the course of three years
}
\keyword{datasets}
