% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor.predictions.R
\name{anchor.predictions}
\alias{anchor.predictions}
\title{Generate Predictions for a Model}
\usage{
anchor.predictions(model, reference, shutup = F)
}
\arguments{
\item{model}{The fitted model object to generate predictions}

\item{reference}{A vector (or single value) containing the name of the variable(s) the user wishes to generate predictions for. For
categorical variables, the function will generate predictions for every level of the categorical variable. For numeric variables, the function
will generate predictions at +1 and -1 standard deviations away from the mean.}

\item{shutup}{The function will give a notice if you don't specify predictions for variables that are in the model. This argument tell it to shut up.}
}
\value{
A data frame containing the predicted values (along with the values of the predictor variables used to estimate the prediction)
}
\description{
Generate Predictions for a Model
}
\details{
This function is simply an easy-to-use wrapper for the predict function. With some models (e.g., logistic regression), the metrics are not very intuitive. 
The anchor.predictions model generates the actual predicted values, depending on what the user specifies.
}
\examples{
data(exercise_data)
linear.model = lm(weight.loss~health + gender, data= exercise_data)
# generate predictions for males/females
anchor.predictions(linear.model, "gender")
# generate predictions for health (+/- 1 sd from the mean)
anchor.predictions(linear.model, "health")
# fit a logistic regression model
data(tablesaw.injury)
glm.mod = glm(injury~safety + attention + gender, data= tablesaw.injury, family=binomial)
anchor.predictions(glm.mod, "attention")
anchor.predictions(glm.mod, c("safety", "gender"))
}
\author{
Dustin Fife
}
