% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{attractiveness}
\alias{attractiveness}
\title{Simulated Dataset About People's Attractiveness}
\format{A data frame with 207 rows and 5 variables: \describe{
  \item{stats_skills}{performance on a statistics know-how test}
  \item{group}{Treatment group (control, watching hitch, receiving stats training)}
  \item{gender}{Gender of person being rated}
  \item{exercise}{Self-reported average number of minutes exercised per week}
  \item{attractiveness_1}{Attractiveness score of person being rated at baseline}
  \item{attractiveness_2}{Attractiveness score of person being rated at follow up}
  \item{age}{Age of respondent}
  \item{gender}{Gender of respondent}
  \item{political}{Political Ideology of respondent}
}}
\usage{
attractiveness
}
\description{
A dataset containing 207 observations, containing the following information
}
\keyword{datasets}
