% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{authors}
\alias{authors}
\title{Sales of books on Amazon}
\format{A data frame with 197,197 rows and 22 variables:
\describe{
\item{Sold.by}{Publisher of book}
\item{Title}{Deidentified book title}
\item{Author}{Deidentified book author}
\item{Kindle.eBooks.Sales.Rank}{Sales rank}
\item{Sale.Price}{Price of book}
\item{Total.Reviews}{Number of reviews}
\item{Average.Rating}{average of reviews}
\item{Preorder}{Available for pre-order?}
\item{Date.Published}{Date published}
\item{Categories}{Fiction category}
\item{DRM}{Digital Rights Management}
\item{KU}{Kindle Unlimited}
\item{Has.ISBN}{Does the book have an ISBN?}
\item{Price.Was.Set.By.Publisher}{Was the price set by the publisher?}
\item{Consolidated.Digital.List.Price}{Dunno}
\item{Daily.Units.Sold}{Estimated number of books sold a day}
\item{Daily.Gross.Sales}{Estimated daily sales}
\item{Daily.Author.Revenue}{Daily author revenue (estimated)}
\item{Publisher}{Publisher type}
\item{Year}{Year published}
\item{Month}{Month Published}
\item{age}{Age of book}
}}
\usage{
authors
}
\description{
A data-scraped dataset of book sales from about 200K authors
}
\keyword{datasets}
