% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{avengers}
\alias{avengers}
\title{Simulated Statistics on the Final Avengers Battle}
\format{
A data frame with 812 rows and 13 variables:
\describe{
  \item{iq}{Intelligence}
  \item{agility}{weighted scores on an obstacle course}
  \item{speed}{Speed in running the 40 meter dash}
  \item{strength}{Pounds lifted in a benchpress}
  \item{damage.resistence}{Amount of skin deflection (in mm) when slapped with a frozen fish}
  \item{flexibility}{Number of inches past their toes they can reach}
  \item{willpower}{Length of time they wait at a DMV for a driver's license}
  \item{superpower}{Does this person have a superpower?}   
  \item{died}{Whether the person died at the final battle}
  \item{kills}{Number of enemies killed in the final battle}
  \item{injuries}{Number of injuries sustained. Anything more than 5 is marked as a 5}
  \item{minutes.fighting}{Number of minutes spent fighting before dying or giving up}
  \item{shots.taken}{Number of shots (with a gun) or punches thrown at an enemy}
}
}
\usage{
avengers
}
\description{
A dataset containing combat attributes of almost 812 
fighters in the final Avengers battle
}
\keyword{datasets}
