% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.bic.R
\name{bf.bic}
\alias{bf.bic}
\title{Compute Bayes Factor from BICs}
\usage{
bf.bic(model1, model2, invert = F)
}
\arguments{
\item{model1}{The first model to be compared}

\item{model2}{The second model to be compared}

\item{invert}{Should the BF be inverted?}
}
\value{
The approximate Bayes Factor of the model comparison
}
\description{
Compute Bayes Factor from BICs
}
\details{
Given two models, the bf.bic function calculates the approximate bayes factor from the two models' BIC values
}
\author{
Dustin Fife
}
