% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{birthweight}
\alias{birthweight}
\title{Dataset seeks to examine what can affect birthweight.}
\format{A data frame with xx rows and 16 variables:
\describe{
  \item{id}{observation id}
	 \item{headcirumference}{Circumference of baby's head.}
  \item{length}{description}
  \item{Birthweight}{Weight at time of baby's head.}
  \item{Gestation}{Length of time baby was gestated.}
  \item{smoker}{is the mother a smoker (1=yes)}
  \item{motherage}{Age of mother.}
  \item{mnocig}{number of cigarettes smoked by mother}
  \item{mheight}{mother's height}
  \item{mppwt}{mother pre pregnancy weight}
  \item{fage}{father's age}
  \item{fedyrs}{father's years of education}
  \item{fnocig}{father number of cigarettes}
  \item{fheight}{father height}
  \item{lowbwt}{binary indicator of whether it's considered low weight or not}
  \item{mage35}{mother's age, centered on 35}
  \item{LowBirthWeight}{Same as above, but labeled}
}}
\usage{
birthweight
}
\description{
This dataset examines how various variables, especially related to parents, can affect the birthweight of a baby.
}
\keyword{datasets}
