% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{blood_pressure}
\alias{blood_pressure}
\title{Dataset examines what affects blood pressure.}
\format{A data frame with xx rows and 4 variables:
\describe{
  \item{BP}{Blood pressure level}
  \item{Exercise}{description}
  \item{Vegetarian}{Whether the subject is vegetarian or not.}
  \item{Stress}{Level of stress}
}}
\usage{
blood_pressure
}
\description{
This dataset looks at how various lifestyle choices affect blood pressure.
}
\keyword{datasets}
