% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cholesterol}
\alias{cholesterol}
\title{Dataset examines what affects cholesterol.}
\format{A data frame with xx rows and 4 variables:
\describe{
  \item{ID}{description}
  \item{Margarine}{description}
  \item{Week}{description}
  \item{Cholesterol}{description}
}}
\usage{
cholesterol
}
\description{
This dataset looks at how margarine effect cholesterol
}
\keyword{datasets}
