% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cas.R
\name{cluster_adjusted_scatter}
\alias{cluster_adjusted_scatter}
\title{Cluster adjusted scatterplot}
\usage{
cluster_adjusted_scatter(
  formula,
  object,
  adjust = c("all", "intercept", "none"),
  ...
)
}
\arguments{
\item{formula}{a flexplot-like formula}

\item{object}{a fitted lme4 object}

\item{adjust}{how the outcome should be adjusted}

\item{...}{any other parameters passed to flexplot}
}
\value{
a cluster-adjusted scatterplot
}
\description{
Cluster adjusted scatterplot
}
\examples{
mod1 = lme4::lmer(y~x + (x | id), data=small_mixed)
cluster_adjusted_scatter(y~x, object = mod1)
}
