% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{criminal_data}
\alias{criminal_data}
\title{Criminal dataset}
\format{
A data frame with xx rows and 6 variables:
\describe{
  \item{ses}{Socioeconomic status of the subject}
  \item{empathy}{description}
  \item{depression}{Whether the subject is depressed}
  \item{rape}{description}
  \item{convictions}{How many times the subject has been convicted of a crime}
  \item{aggression}{How aggressive the subject is}
}
}
\usage{
criminal_data
}
\description{
This is a simulated dataset that looks at how various variables affect convictions and rape instances
}
\keyword{datasets}
