% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{diet}
\alias{diet}
\title{Diet dataset}
\format{A data frame with xx rows and 6 variables:
\describe{
  \item{Person}{description}
  \item{gender}{The subject's gender}
  \item{Age}{Age of the subject}
  \item{Height}{Height of the subject}
  \item{pre.weight}{Weight of the subject before the diet}
  \item{weight6weeks}{Weight of the subject after 6 weeks}
  \item{Diet}{Type of diet the subject went on}
}}
\usage{
diet
}
\description{
This dataset seeks to examine how effective certain diets can be for losing weight.
}
\keyword{datasets}
