% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{drug_use}
\alias{drug_use}
\title{Drug use and SES}
\format{
A data frame with xx rows and 3 variables:
\describe{
  \item{drug.use}{description}
  \item{SES}{Socioeconomic status of the subject}
  \item{therapy}{Whether subjects attends therapy for drug use}
}
}
\usage{
drug_use
}
\description{
This dataset looks at a possible relationship between drug usage and socioeconomic status.
}
\keyword{datasets}
