% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.glm}
\alias{estimates.glm}
\alias{estimates}
\title{Report glm object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{glm}(object)
}
\arguments{
\item{object}{a glm object}
}
\description{
Report glm object Estimates
}
