% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_glmer.R
\name{estimates.glmerMod}
\alias{estimates.glmerMod}
\title{Report glmerMod object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{glmerMod}(object, mc = FALSE)
}
\arguments{
\item{object}{a glmerMod object}

\item{mc}{Should model comparisons be performed? Currently not used}
}
\value{
One or more objects containing parameter estimates and effect sizes
}
\description{
Report glmerMod object Estimates
}
