% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{estimates.lmerMod}
\alias{estimates.lmerMod}
\title{Report lmerMod object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{lmerMod}(object, mc = TRUE)
}
\arguments{
\item{object}{a lm object}

\item{mc}{Should model comparisons be performed?}
}
\value{
One or more objects containing parameter estimates and effect sizes
}
\description{
Report lmerMod object Estimates
}
