% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.RandomForest}
\alias{estimates.RandomForest}
\title{Report RandomForest object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{RandomForest}(object)
}
\arguments{
\item{object}{a RandomForest object}
}
\value{
One or more objects containing parameter estimates and effect sizes
}
\description{
Report RandomForest object Estimates
}
