% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{estimates}
\alias{estimates}
\title{Report Estimates (effect sizes and parameters)}
\usage{
estimates(object, mc = TRUE)
}
\arguments{
\item{object}{a object}

\item{mc}{Should model comparisons be performed?}
}
\description{
Report object Estimates
}
