% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.regression}
\alias{estimates.regression}
\alias{estimates}
\title{Report regression object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{regression}(object)
}
\arguments{
\item{object}{a regression object}
}
\description{
Report regression object Estimates
}
