% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.zeroinfl}
\alias{estimates.zeroinfl}
\alias{estimates}
\title{Report zeroinfl object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{zeroinfl}(object)
}
\arguments{
\item{object}{a zeroinfl object}
}
\description{
Report zeroinfl object Estimates
}
