% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{exercise_data}
\alias{exercise_data}
\title{Exercise dataset}
\format{A data frame with xx rows and 13 variables:
\describe{
  \item{therapy.type}{description}
  \item{health}{description}
  \item{motivation}{description}
  \item{weight.loss}{description}
  \item{gender}{description}
  \item{satisfaction}{description}
  \item{rewards}{description}
  \item{income}{description}
  \item{muscle.gain}{description}
  \item{muscle.gain.missing}{description}
  \item{six.mo.weight}{description}
}}
\usage{
exercise_data
}
\description{
Simulated data that sees how weight loss varies as a function of various predictors, and treatment condition
}
\keyword{datasets}
