% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexplot.R
\name{flexplot}
\alias{flexplot}
\title{Create a "flexible plot" or flexplot}
\usage{
flexplot(formula, data, related = F, color = NULL, symbol = NULL,
  linetype = NULL, bins = 4, labels = NULL, breaks = NULL,
  method = "loess", se = T, ghost.line = NULL,
  ghost.reference = NULL, spread = c("quartiles", "stdev", "sterr"),
  jitter = FALSE, raw.data = T, sample = Inf, prediction = NULL,
  suppress_smooth = F, alpha = 0.99977)
}
\arguments{
\item{formula}{A formula of the form  y~x + a | b + z}

\item{data}{The dataset}

\item{related}{Are variables related? If so, it will show a plot of difference scores, rather than the raw scores (for a related t test).}

\item{color}{Color values to be used for the categorical variables}

\item{symbol}{The symbols to be used for the categorical variables}

\item{linetype}{The linetype to be used for the categorical variables}

\item{bins}{The number of bins used when putting quantitative variables into categories. A list can be used if different amounts are wanted for different variables}

\item{labels}{A list of the same length as bins}

\item{breaks}{The breaks to be used for the bins}

\item{method}{The method to be used to draw the lines. Defaults to loess}

\item{se}{Should standard errors be drawn?}

\item{ghost.line}{Should a ghost line be drawn? If so, user must specify a color. Default is NULL (in which case, the ghost line isn't shown).}

\item{spread}{How should standard errors be drawn? Defaults to quartiles}

\item{jitter}{Should values be jittered?}

\item{raw.data}{Should raw data be plotted?}

\item{sample}{Should a sample of the data be plotted? Defaults to Inf (for all variables).}

\item{ghost.group}{What should the reference group be (from which to draw the ghost line)? The user can specify up to two values as a vector. See examples.}
}
\description{
Create a flexible plot
}
\details{
Formula takes the form of y~x + a | b + z. Everything on the right of | will occur in panels (in this case, b will be in the columns and z will be in the rows). Any numeric
variables between ~ and | will fall on the x axis and categorical predictors will be color-coded/symbol-coded/line-coded (if user specifies). If more than one numeric variable is specified
after ~, the second (and third) will be binned. No more than four variables are allow (otherwise, cognitive load is too high). If the user wishes to include another variable,
 I recommend they create separate plots, one for each level of the new variable (or a binned level of a numeric variable).
}
\examples{
clear()
data(exercise_data)
d = exercise_data

# # #### histograms and barcharts
flexplot(income~1, data=d)
flexplot(gender~1, data=d)

# ### scatter plot
source("research/RPackages/fifer/R/flexplot.R")
flexplot(weight.change~motivation, data=d)	
flexplot(weight.change~motivation, data=d, method="lm", se=F)	#### with regression line and without standard error	

# ### mean plots
flexplot(weight.change~therapy.type, data=d)
flexplot(weight.change~therapy.type, data=d, raw.data=F)		## without raw data

# ### CHI SQUARE PLOT
flexplot(therapy.type~gender, data=d)	
		
# ### INTERACTION PLOT			
flexplot(weight.change~therapy.type + gender, data=d)
flexplot(weight.change~therapy.type + gender, data=d, sample=50)	#### sampling 50 people instead (to make it less noisy)

# #### ANCOVA PLOT
flexplot(weight.change~motivation + gender, data=d, se=F)	### remove se

# #### 2N PLOT (2 NUMERIC VARIABLE PLOTS)
flexplot(weight.change~motivation + income, data=d, se=F, method="lm")
flexplot(weight.change~motivation + income, data=d, se=F, method="lm", breaks = list(c(95000, 100000, 105000)),labels=list(c("<95K", "<100K", "<105K", ">105K")))		### change labels for income

# #### 3N plot
flexplot(weight.change~motivation + income + health, data=d, se=F, method="lm")	## different lines for income
flexplot(weight.change~motivation | income + health, data=d, se=F, method="lm")	## different panels for income
flexplot(weight.change~motivation | income + health, data=d, se=F, method="lm", breaks = list(c(95000, 100000, 105000)),labels=list(c("<95K", "<100K", "<105K", ">105K")))	## relabel income

# #### three categorical variables (multiway dot plot)
flexplot(weight.change~gender + therapy.type + rewards, data=d, raw.data=F)
}
\author{
Dustin Fife
}
