% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexplot_jasp2.R
\name{flexplot_jasp2}
\alias{flexplot_jasp2}
\title{Flexplot in JASP}
\usage{
flexplot_jasp2(jaspResults, dataset, options)
}
\arguments{
\item{jaspResults}{A JASP object}

\item{dataset}{dataset supplied by JASP}

\item{options}{a list of options to pass to JASP}
}
\value{
a flexplot graphic.
}
\description{
This function was developed for use in JASP. It takes a dataset as input with user 
options and returns a flexplot graphic
}
