% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexplota.h.R
\name{flexplota}
\alias{flexplota}
\title{Flexplot}
\usage{
flexplota(data, out, preds, given, se = TRUE, line = "Loess",
  ghost = TRUE, resid = FALSE, center = "Median + quartiles",
  alpha = 50)
}
\arguments{
\item{data}{.}

\item{out}{.}

\item{preds}{.}

\item{given}{.}

\item{se}{.}

\item{line}{.}

\item{ghost}{.}

\item{resid}{.}

\item{center}{.}

\item{alpha}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Flexplot
}
