% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{floor_ceiling}
\alias{floor_ceiling}
\title{Replace minimum value with a specified floor}
\usage{
floor_ceiling(x, min.val = NULL, max.val = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{min.val}{The minimum value of the dataset}

\item{max.val}{The maximum value of the dataset}
}
\value{
A new vector where the values < min.val are replaced with min.val (and
likewise for the max values)
}
\description{
Replace minimum value with a specified floor
}
\examples{
testval = c(-1, 0, 0, 60, 100, 101)
floor_ceiling(testval, min.val=-1, max.val=100)
}
