% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_compare_fits.R
\name{get_model_n}
\alias{get_model_n}
\title{Get Number of Observations from Model Objects}
\usage{
get_model_n(model)
}
\arguments{
\item{model}{A fitted model object}
}
\value{
Integer representing the number of observations, or NULL if unknown
}
\description{
Extract the number of observations used to fit a model object.
}
