% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_compare_fits.R
\name{get_terms}
\alias{get_terms}
\title{Extract Variable Names from Model Objects}
\usage{
get_terms(model)
}
\arguments{
\item{model}{A fitted model object (e.g., lm, glm, RandomForest, keras model, etc.)}
}
\value{
A list with two elements:
  \item{predictors}{Character vector of predictor variable names}
  \item{response}{Character string of the response variable name}
}
\description{
This function extracts predictor and response variable names from various
types of fitted model objects. It is used internally by flexplot functions
to determine which variables are available in a model.
}
\details{
This generic function provides a consistent interface for extracting variable
names across different model types. Each model class should have its own method
that handles the specifics of variable extraction for that model type.
}
\examples{
# Linear model example
data(mtcars)
model = lm(mpg ~ hp + wt, data = mtcars)
terms_info = get_terms(model)
terms_info$predictors  # "hp" "wt"
terms_info$response    # "mpg"

}
\seealso{
\code{\link{compare.fits}} for the main function that uses this generic
}
