% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmbasic.h.R
\name{glmbasic}
\alias{glmbasic}
\title{General Linear Model}
\usage{
glmbasic(data, out, preds, graphic = TRUE, graphicassump = TRUE,
  estimates = TRUE, se = TRUE, line = "loess",
  center = "median + quartiles")
}
\arguments{
\item{data}{.}

\item{out}{.}

\item{preds}{.}

\item{graphic}{.}

\item{graphicassump}{.}

\item{estimates}{.}

\item{se}{.}

\item{line}{.}

\item{center}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$assumpplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$glmcat} \tab \tab \tab \tab \tab a table \cr
  \code{results$rsq} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$glmcat$asDF}

\code{as.data.frame(results$glmcat)}
}
\description{
General Linear Model
}
