% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{graduate_income}
\alias{graduate_income}
\title{Graduate income}
\format{A data frame with xx rows and 4 variables:
\describe{
  \item{Grad.School}{Whether the subject attended grad school}
  \item{Profession}{Profession of the subject}
  \item{GPA}{GPA of the subject}
  \item{Income}{The current income of the subject}
  \item{Years}{Years since they graduated}
}}
\usage{
graduate_income
}
\description{
This dataset examines whether schools and academic performance can lead to higher income in professions. Simulated data
}
\keyword{datasets}
