% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimates.R, R/icc.R
\name{icc}
\alias{icc}
\title{Compute the ICC from a lmer (package lme4) model}
\usage{
icc(model)

icc(model)
}
\arguments{
\item{model}{a lmer model}
}
\value{
the ICC

the ICC
}
\description{
This function will extract the variances from a mixed model and output the value of the ICC

This function will extract the variances from a mixed model and output the value of the ICC
}
\details{
Recall that the equation for the ICC is var(school)/[var(school)+ var(person)]. 
This forumla extracts the necessary variances from the mixed model and computes it.

Recall that the equation for the ICC is var(school)/[var(school)+ var(person)]. 
This forumla extracts the necessary variances from the mixed model and computes it.
}
\examples{
data(alcuse)
require(lme4)
model = lmer(ALCUSE~1|ID, data=alcuse)
icc(model)
data(alcuse)
require(lme4)
model = lmer(ALCUSE~1|ID, data=alcuse)
icc(model)
}
\author{
Dustin Fife

Dustin Fife
}
