% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ideation}
\alias{ideation}
\title{Simulated Dataset About Suicide Ideation}
\format{
A data frame with 300 rows and 8 variables: \describe{
  \item{ideation}{Suicide ideation scores}
  \item{stress}{Self-reported stress score}
  \item{health}{Self-reported health score}
  \item{friend_ideation}{Suicide ideation score of each individual's best friend}
  \item{depression_c}{Centered BDI depression score}
  \item{friend_ideation_c}{Centered friend ideation score}
  \item{stress}{centered stress score}   
}
}
\usage{
ideation
}
\description{
A dataset containing 3000 observations. This dataset contains interaction and nonlinear effects
and was used for the manuscript about Visual partitioning
}
\keyword{datasets}
