% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{job_data}
\alias{job_data}
\title{Job performance dataset}
\format{A data frame with 187 rows and 4 variables:
\describe{
  \item{job.performance}{supervisor ratings of job performance}
  \item{iq}{intelligence}
  \item{conscientiousness}{score on a conscientious scale}
  \item{interview}{Rating on an interview scale}
}}
\usage{
job_data
}
\description{
These simulated data show relationships between various predictors of job performance with missing data
}
\keyword{datasets}
