% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_avp.R
\name{logistic_added_plot}
\alias{logistic_added_plot}
\title{Create an added variable plot with binned residuals for logistic regression}
\usage{
logistic_added_plot(
  formula,
  data,
  lm_formula = NULL,
  method = "loess",
  x = NULL,
  n_bins = 10,
  scale = "probability",
  ...
)
}
\arguments{
\item{formula}{A flexplot formula, specifying how the avp will visualize the variables.}

\item{data}{The dataset used}

\item{lm_formula}{Optional. A formula specifying how to condition variables (the logistic model to fit first).}

\item{method}{The smoothing method. Defaults to "loess"}

\item{x}{The variable you wish to place on the x axis. Defaults to NULL.}

\item{n_bins}{Number of bins to use for calculating binned residuals. Defaults to 10.}

\item{scale}{Either "probability" (default) or "logit" to determine the scale of residuals.}

\item{...}{Other parameters passed to flexplot}
}
\value{
An added variable plot with binned residuals
}
\description{
Create an added variable plot using binned percentage residuals instead of raw residuals.
This function first fits a logistic regression model to residualize the outcome variable 
based on conditioning variables, then bins the remaining predictor variable and calculates
binned percentage residuals (observed proportion - fitted proportion within each bin).
}
\details{
The function works similarly to added.plot but uses the binning logic from logistic_residual_plots
to create residuals that are more appropriate for logistic regression diagnostics.
}
\examples{
# Assuming you have a binary outcome
# logistic_added_plot(outcome~predictor1 + predictor2, data=mydata)
# logistic_added_plot(outcome~predictor1 + predictor2, data=mydata, x=2)
# logistic_added_plot(outcome~predictor1 + predictor2, 
#      lm_formula = outcome~confounder1*confounder2, data=mydata)
# logistic_added_plot(outcome~predictor1 + predictor2, data=mydata, scale="logit")
}
\seealso{
\code{\link{added.plot}}, \code{\link{logistic_residual_plots}}
}
\author{
Dustin Fife
}
