% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_estimates.R
\name{logistic_coefficients}
\alias{logistic_coefficients}
\title{Output calculus-based logistic regression coefficients}
\usage{
logistic_coefficients(model, location = 0.5)
}
\arguments{
\item{model}{A fitted model (glm or glmerMod)}
}
\value{
Discrimination and difficulty estimates for a regression model
}
\description{
This function takes a logistic regression model and returns more intuitive slope/intercept
parameters. The slope (discrimination) is interpreted as the instantaneous (tangent) slope of the ogive
curve at p =  0.5. The intercept (difficulty) is interpreted as the value of X where
the logistic curve crosses the 50% threshold
}
\examples{
logistic_model = glm(y_bin~x + a, data=small, family=binomial)
logistic_coefficients(logistic_model)
}
