% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_plot.R
\name{marginal_plot}
\alias{marginal_plot}
\title{Create "margin plots"}
\usage{
marginal_plot(p, columns = TRUE, rows = TRUE, grand_mean = TRUE)
}
\arguments{
\item{p}{A flexplot graphic that has paneled variables.}

\item{columns}{Logical. Should column panels be plotted?}

\item{rows}{Logical. Should row panels be plotted?}

\item{grand_mean}{Logical. Stating TRUE will generate a simple bivariate plot between X and Y}
}
\value{
A graphic with marginal plots
}
\description{
This function takes as an argument a regular flexplot graphic that has panels, then plots the "marginal" relationships for the panels. 
This makes it easier to see the average effect across the rows (or columns).
}
\examples{
p = flexplot(weight.loss~motivation | gender + rewards, data=exercise_data)
marginal_plot(p)
marginal_plot(p, rows=FALSE)
}
