% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{math}
\alias{math}
\title{Math Achievement Dataset}
\format{
A data frame with xx rows and 6 variables:
\describe{
  \item{School}{Which school the subject attends}
  \item{Minority}{description}
  \item{Sex}{Gender of the subject}
  \item{SES}{Socioeconomic status of the subject}
  \item{MathAch}{description}
  \item{MEANSES}{description}
}
}
\usage{
math
}
\description{
Examines the relationship between various variables and math achievement. Great for HLM.
}
\keyword{datasets}
