% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/added.plot.R
\name{mediate_plot}
\alias{mediate_plot}
\title{Create a plot for mediation analysis}
\usage{
mediate_plot(formula, data, method = "lm", ...)
}
\arguments{
\item{formula}{A formula with y on the left side and the predictors on the right. The last variable
entered will be plotted on the x-axis}

\item{data}{The dataset}

\item{method}{The fitted line. Defaults to "lm".}

\item{...}{Other parameters passed to flexplot}
}
\value{
a plot showing the mediation effect
}
\description{
Create a plot for mediation analysis
}
\examples{
mediate_plot(weight.loss~motivation + health, data=exercise_data)
mediate_plot(weight.loss~motivation + therapy.type, data=exercise_data)
}
