% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.mod.visual.R
\name{mixed.mod.visual}
\alias{mixed.mod.visual}
\title{Display a graphic of a mixed model}
\usage{
mixed.mod.visual(formula, data, model, n = 6, jitter = F)
}
\arguments{
\item{formula}{A flexplot friendly formula, though it can only include one predictor}

\item{data}{The dataset}

\item{model}{The model}

\item{n}{The number of IDs to sample and put in different panels}

\item{jitter}{Should the data be jittered?}
}
\value{
A plot
}
\description{
Display a graphic of a mixed model
}
\details{
This function samples the ID variable and plots the bivariate relationship, depending on the mixed model chosen
}
\author{
Dustin Fife
}
