% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedmod_jasp.R
\name{mixedmod_jasp}
\alias{mixedmod_jasp}
\title{General Linear Model in JASP}
\usage{
mixedmod_jasp(jaspResults, dataset, options)
}
\arguments{
\item{jaspResults}{A JASP object}

\item{dataset}{dataset supplied by JASP}

\item{options}{a list of options to pass to JASP}
}
\value{
a table, plot, etc.
}
\description{
This function was developed for use in JASP. It takes a dataset as input with user 
options and returns results (tables, plots, etc)
}
