% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comparison.R
\name{model.comparison}
\alias{model.comparison}
\title{Perform a model comparison}
\usage{
model.comparison(model1, model2)
}
\arguments{
\item{model1}{the first model to be compared}

\item{model2}{the second model to be compared}
}
\description{
Perform a model comparison
}
\details{
This function takes two models (either nested or not nested) and compares them in terms of AIC, BIC,
bayes factor, and the p-value
}
\author{
Dustin Fife
}
