% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomp.h.R
\name{modelcomp}
\alias{modelcomp}
\title{Model Comparison}
\usage{
modelcomp(data, dep, pred, pred2)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{pred}{.}

\item{pred2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Model Comparison
}
