% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modifiers.R
\name{modify_labels}
\alias{modify_labels}
\title{Modify the labels in a flexplot (or ggplot) graphic}
\usage{
modify_labels(
  p,
  x = NULL,
  y = NULL,
  colors = NULL,
  panel_rows = NULL,
  panel_cols = NULL
)
}
\arguments{
\item{p}{The flexplot (or ggplot) object}

\item{x}{The label for the x axis}

\item{y}{The label for the y axis}

\item{colors}{The label for the variable encoded as a color}

\item{panel_rows}{The label for the variable encoded in row panels}

\item{panel_cols}{The label for the variable encoded in column panels}
}
\value{
A new plot with the modifications
}
\description{
Modify the labels in a flexplot (or ggplot) graphic
}
