% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_points.R
\name{modify_points}
\alias{modify_points}
\title{Modify the points in a flexplot (or ggplot) graphic}
\usage{
modify_points(p, shape = 19, colour = "black", size = 1.5)
}
\arguments{
\item{p}{The flexplot (or ggplot) object}

\item{shape}{The new shape desired. Defaults to 19 (circles).}

\item{colour}{The new colour desired. Defaults to "black".}

\item{size}{The new size of the points desired. Defaults to 1.5.}
}
\value{
A new plot with the modifications
}
\description{
Modify the points in a flexplot (or ggplot) graphic
}
