% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modifiers.R
\name{modify_smooth}
\alias{modify_smooth}
\title{Modify the fitted line in a flexplot (or ggplot) graphic}
\usage{
modify_smooth(
  p,
  color = NULL,
  size = NULL,
  linetype = NULL,
  method = NULL,
  se = NULL
)
}
\arguments{
\item{p}{The flexplot (or ggplot) object}

\item{color}{The color of the line}

\item{size}{The size (thickness) of the line}

\item{linetype}{The type of line as a string (e.g., "dotted")}

\item{method}{The type of smoothing function applied. Can be lm, quadratic, cubic, rlm, logistic,
poisson, or loess}

\item{se}{Standard errors. If TRUE, a band will be displayed showing the standard errors}
}
\value{
A new plot with the modifications
}
\description{
Modify the fitted line in a flexplot (or ggplot) graphic
}
