% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{nsduh}
\alias{nsduh}
\title{NSDUH Dataset}
\format{A data frame with xx rows and 27 variables:
\describe{
  \item{cig.rec}{smoked cigarettes recently?}
  \item{alc.rec}{used alcohol recently?}
  \item{coc.rec}{used cocain recently?}
  \item{her.rec}{used heroin recently?}
  \item{alc.freq}{frequency of alcohol use}
  \item{cocain.freq}{frequency of cocain use}
  \item{heroin.freq}{frequency of heroin use}
  \item{cig.freq}{frequency of cigarette use}
  \item{inpatient}{have they received inpatient treatment?}
  \item{outpatient}{have they received outpatient treatment?}
  \item{moves.past5}{moved in the past 5 years?}
  \item{attend.rel.serv}{do you attend religious services?}
  \item{rel.inf.dec}{how influential is religion on your decisions?}
  \item{distress}{mental distress score}
  \item{whodas.impairment}{mental impairment score}
  \item{MI}{probability of mental illness}
  \item{major.dep}{ever had major depression}
  \item{dep.pas.yr}{depression in the past year}
  \item{age}{age}
  \item{military}{been in military}
  \item{health.rating}{self reported health rating}
  \item{sex}{sex}
  \item{education}{educational attainment}
  \item{race}{race}
  \item{county}{type of county they live in (small metro, large metro, etc)}
  \item{income}{how much bank they make}
}}
\usage{
nsduh
}
\description{
This is the 2014 National Survey of Drug Use and Health (NSDUH) dataset
}
\keyword{datasets}
