% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{photo_lessons}
\alias{photo_lessons}
\title{An actual dataset of photography students}
\format{
A tibble with 162 rows and 7 variables: 
\describe{
  \item{clientid}{A unique identifier for each student taking a class} 
  \item{link_name}{The name of the class each student took}
  \item{class_size}{The number of students in that specific class}
  \item{classes_taken}{The total number of classes each student has taken}
  \item{redroom}{Whether the student took class in a very large, dark 
  room called the red room}
  \item{dates_between}{The number of days between their first class and their next class}
  \item{progre3ssed}{A Binary variable indicating whether the student decided 
  to take advanced classes}
  \item{class_id}{A unique identifer for the class}
  }
}
\usage{
photo_lessons
}
\description{
This dataset contains 162 rows of repeated measures data for photography 
students across 8 variables. These data were recorded from 2008-2012 
from a photography-training school.
}
\keyword{datasets}
