% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{plant_growth}
\alias{plant_growth}
\title{Plant Growth}
\format{A data frame with xx rows and 5 variables:
\describe{
  \item{Location}{Location stored (greenhouse vs inside)}
  \item{Soil.Type}{Type of soil used}
  \item{Method}{Method of planting (soil blocks versus newspaper)}
  \item{Diameter}{Diameter of trunk}
  \item{Leaves}{Number of leaves}
}}
\usage{
plant_growth
}
\description{
Measures of seedling growth depending on various characteristics. NOT simulated data.
}
\keyword{datasets}
