% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jitterd.R
\name{position_jitterd}
\alias{position_jitterd}
\title{Jitter points based on density to avoid overplotting}
\usage{
position_jitterd(width = NULL, height = NULL, quad.points = 100,
  seed = NA)
}
\arguments{
\item{width, height}{Amount of vertical and horizontal jitter. The jitter
  is added in both positive and negative directions, so the total spread
  is twice the value specified here. However, that amount of jittering depends
     on the density. The location of highest density in the dataset will have 
  jittering equal to the amount the user specifies, otherwise, jittering is a 
  fraction of that. 

  If omitted, defaults to 40\% of the resolution of the data: this means the
  jitter values will occupy 80\% of the implied bins for the highest density. Categorical data
  is aligned on the integers, so a width or height of 0.5 will spread the
  data so it's not possible to see the distinction between the categories.}

\item{seed}{A random seed to make the jitter reproducible.
  Useful if you need to apply the same jitter twice, e.g., for a point and
  a corresponding label.
  The random seed is reset after jittering.
  If `NA` (the default value), the seed is initialised with a random value;
  this makes sure that two subsequent calls start with a different seed.
  Use `NULL` to use the current random seed and also avoid resetting
  (the behaviour of \pkg{ggplot} 2.2.1 and earlier).}
}
\description{
Counterintuitively adding random noise to a plot can sometimes make it
easier to read. Jittering is particularly useful for small datasets with
at least one discrete position. Jittering based on density ensures that
the amount of random noise only increases as the probability of overlap
increases.
}
\examples{
# Jittering is useful when you have a discrete position, and a relatively
# small number of points
# take up as much space as a boxplot or a bar
ggplot(mpg, aes(class, hwy)) +
  geom_boxplot(colour = "grey50") +
  geom_jitterd()

# If the default jittering is too much, as in this plot:
ggplot(mtcars, aes(am, vs)) +
  geom_jitterd()

# You can adjust it in two ways
ggplot(mtcars, aes(am, vs)) +
  geom_jitterd(width = 0.1, height = 0.1)
ggplot(mtcars, aes(am, vs)) +
  geom_jitterd(position = position_jitter(width = 0.1, height = 0.1))

# Create a jitter object for reproducible jitter:
jitter <- position_jitterd(width = 0.1, height = 0.1)
ggplot(mtcars, aes(am, vs)) +
  geom_point(position = jitterd) +
  geom_point(position = jitterd, color = "red", aes(am + 0.2, vs + 0.2))
}
\concept{position adjustments}
