% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jitterd.R
\name{position_jitterdodged}
\alias{position_jitterdodged}
\title{Simultaneously dodge and jitter}
\usage{
position_jitterdodged(jitter.width = 0.2, jitter.height = 0,
  dodge.width = 0.75, seed = NA)
}
\arguments{
\item{jitter.width}{degree of jitter in x direction. Defaults to 40\% of the
resolution of the data.}

\item{jitter.height}{degree of jitter in y direction. Defaults to 0.}
}
\description{
This is primarily used for aligning points generated through
`geom_point()` with dodged boxplots (e.g., a `geom_boxplot()` with
a fill aesthetic supplied).
}
\examples{
dsub <- diamonds[ sample(nrow(diamonds), 1000), ]
ggplot(dsub, aes(x = cut, y = carat, fill = clarity)) +
  geom_boxplot(outlier.size = 0) +
  geom_point(pch = 21, position = position_jitterdodged())
}
\seealso{
Other position adjustments: \code{\link{position_jitterd}}
}
\concept{position adjustments}
