% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_helper.R, R/print_estimates_helper.R
\name{print.estimates}
\alias{print.estimates}
\title{Print estimates Summary}
\usage{
\method{print}{estimates}(x, ...)

\method{print}{estimates}(x, ...)
}
\arguments{
\item{x}{an estimates object}

\item{...}{ignored}
}
\value{
A printed list of estimates

A printed list of estimates
}
\description{
Print estimates Summary

Print estimates Summary
}
