% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_estimates_helper.R
\name{print.glm_estimates}
\alias{print.glm_estimates}
\title{Print glm_estimates Summary}
\usage{
\method{print}{glm_estimates}(x, ...)
}
\arguments{
\item{x}{an glm_estimates object}

\item{...}{ignored}
}
\value{
A printed list of estimates
}
\description{
Print a glm_estimates object
}
