% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_helper.R, R/print_estimates_helper.R
\name{print.glmer_estimates}
\alias{print.glmer_estimates}
\title{Print glmer_estimates Summary}
\usage{
\method{print}{glmer_estimates}(x, ...)

\method{print}{glmer_estimates}(x, ...)
}
\arguments{
\item{x}{an glmer_estimates object}

\item{...}{ignored}
}
\value{
A printed list of estimates

A printed list of estimates
}
\description{
Print a glmer_estimates object

Print a glmer_estimates object
}
