% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{print.rf_estimates}
\alias{print.rf_estimates}
\title{Print rf_estimates Summary}
\usage{
\method{print}{rf_estimates}(x, ...)
}
\arguments{
\item{x}{an rf_estimates object}

\item{...}{ignored}
}
\description{
Print a rf_estimates object
}
