% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{relationship_satisfaction}
\alias{relationship_satisfaction}
\title{Relationship satisfaction dataset}
\format{
A data frame with xx rows and 7 variables:
\describe{
  \item{conscientiousness}{description}
  \item{honesty}{description}
  \item{communication}{description}
  \item{interests}{description}
  \item{satisfaction}{description}
  \item{separated}{description}
  \item{gender}{description}
}
}
\usage{
relationship_satisfaction
}
\description{
Data about various factors affecting relationship satisfaction. Simulated data.
}
\keyword{datasets}
