% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Transform a variable to have a particular mean and standard deivation}
\usage{
rescale(x, new.mean, new.sd)
}
\arguments{
\item{x}{A vector containing the data the user wishes to transform}

\item{new.mean}{The desired mean of the dataset}

\item{new.sd}{The desired sd of the dataset}
}
\value{
a vector of rescaled values
}
\description{
This function will take a variable and transform it such that it has a particular mean and variance
}
\examples{
x = rnorm(100, 0, 1)
## transform to have a mean of 20 and sd of 5
x.new = rescale(x, 50, 5)
mean(x.new); sd(x.new)
}
\author{
Dustin Fife
}
