% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{secret_agent}
\alias{secret_agent}
\title{Secret agent dataset}
\format{A data frame with xx rows and 7 variables:
\describe{
  \item{muscle}{bicep circumference}
  \item{height}{height in inches}
  \item{tortured}{have they been tortured on a mission?}
  \item{missions}{number of missions completed}
  \item{Course.Score}{score on an obstacle course}
  \item{Branch}{Branch of service (CIA versus MI6)}
  \item{Salary}{Annual salary}
}}
\usage{
secret_agent
}
\description{
Simulated data about various characteristics of secret agents
}
\keyword{datasets}
