% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comparison.R
\name{sensitivity.table}
\alias{sensitivity.table}
\title{Compute sensitivity/specificity/etc.}
\usage{
sensitivity.table(object)
}
\arguments{
\item{object}{an object that can be predicted (e.g., glm). Note the thing to be predicted must have only two outcomes}
}
\value{
A table containing sensitivity/specificity/etc.
}
\description{
Compute sensitivity/specificity/etc.
}
\details{
This function computes sensitivity, specificity, positive/negative predictive value and accuracy and reports
them as a list.
}
\author{
Dustin Fife
}
