% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{small}
\alias{small}
\title{Small test dataset}
\format{
A data frame with xx rows and 5 variables:
\describe{
  \item{y}{normal variable}
  \item{a}{categorical variable}
  \item{b}{another categorical variable}
  \item{z}{normally distributed predictor}
  \item{x}{another normally distributed predictor}
  \item{y_bin}{a binary outcome}
  \item{y_pois}{a count outcome}
  \item{y_gam}{a gamma outcome}      
}
}
\usage{
small
}
\description{
Simulated data for testing purposes
}
\keyword{datasets}
