% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized.beta.R
\name{standardized.beta}
\alias{standardized.beta}
\title{Compute standardized betas on a linear model object}
\usage{
standardized.beta(object, sd.y = T, se = F)
}
\arguments{
\item{object}{a lm object}

\item{sd.y}{Should we scale based on standard deviation of Y as well as the predictors?}

\item{se}{Should standard errors be reported?}
}
\value{
the standardized betas
}
\description{
Compute standardized betas on a linear model object
}
\details{
Compute standardized betas on a linear model object
}
\author{
Dustin Fife
}
