% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{tablesaw.injury}
\alias{tablesaw.injury}
\title{Tablesaw injuries}
\format{A data frame with xx rows and 5 variables:
\describe{
  \item{safety}{description}
  \item{attention}{description}
  \item{gender}{description}
  \item{probs}{description}
  \item{injury}{description}
}}
\usage{
tablesaw.injury
}
\description{
Simulated data about the probability of getting injured by a tablesaw
}
\keyword{datasets}
