% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni.plot.R
\name{uni.plot}
\alias{uni.plot}
\title{Plot univariate distribution of a variable}
\usage{
uni.plot(variable, d = NULL, numeric = NULL)
}
\arguments{
\item{variable}{A string indicating the name of the variable to plot}

\item{d}{The name of the dataset (which contains the variable the user wishes to plot)}

\item{numeric}{Logical. Is the variable of interest numeric? (Meaning, should a histogram be plotted?)}
}
\value{
A plot
}
\description{
A function that automatically plots either a barchart or a histogram (depending on data type)
}
\details{
Ideally, the user will provide the data type for this function (by saying numeric = T, or numeric = F)
Omitting that, the function will decide for the user. If the variable is coded as numeric AND there are more than five unique values of the variable,
it will generate a histogram. If the variable is a factor OR it has five or less unique values, it will output a barchart.
}
\examples{
distress = sample(1:10, size=22, replace=T)
uni.plot(distress, numeric=T, d=NULL)
}
\seealso{
\code{\link{MASS}}
}
\author{
Dustin Fife
}
