% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_univariate_lists.R
\name{univariate_list}
\alias{univariate_list}
\title{Produce a bunch of univariate plots}
\usage{
univariate_list(names, data)
}
\arguments{
\item{names}{A vector containing the names of the variables you want to plot}

\item{data}{A dataset}
}
\value{
A list of plots
}
\description{
Produce a bunch of univariate plots
}
\examples{
univariate_lists(c("a", "b", "z"), small)
}
