% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.default}
\alias{visualize.default}
\title{Visualize a fitted model}
\usage{
\method{visualize}{default}(object, plot = c("all", "residuals", "model"), formula = NULL, ...)
}
\arguments{
\item{object}{a model object}

\item{plot}{what should be plotted? Residuals? model plot? All of them?}

\item{formula}{A flexplot-style formula}

\item{...}{Other arguments passed to flexplot}
}
\value{
a plot containing a visual of the chosen model
}
\description{
Visualize a fitted model
}
