% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.glmerMod}
\alias{visualize.glmerMod}
\title{Visualize a fitted glmerMod model}
\usage{
\method{visualize}{glmerMod}(
  object,
  plot = c("all", "residuals", "model"),
  formula = NULL,
  sample = 3,
  plots.as.list = FALSE,
  ...
)
}
\arguments{
\item{object}{a glmer object}

\item{plot}{what should be plotted? Residuals? Model plot? All of them?}

\item{formula}{A flexplot-style formula}

\item{sample}{The number of individuals' datapoints to sample as raw data. Defaults to 3}

\item{plots.as.list}{Should the plots be returned as a list? Defaults to FALSE.}

\item{...}{Other arguments passed to flexplot}
}
\description{
Visualize a fitted glmerMod model
}
