% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.lm}
\alias{visualize.lm}
\title{Visualize a fitted model}
\usage{
\method{visualize}{lm}(object, plot = c("all", "residuals", "model"),
  formula = NULL, ...)
}
\arguments{
\item{object}{a lm object}

\item{plot}{what should be plotted? Residuals? Model plot? All of them?}

\item{formula}{A flexplot-style formula}

\item{...}{Other arguments passed to flexplot}
}
\description{
Visualize a fitted model
}
