% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.RandomForest}
\alias{visualize.RandomForest}
\title{Visualize a fitted RandomForest model}
\usage{
\method{visualize}{RandomForest}(object, plot = c("all", "residuals", "model"), formula = NULL, ...)
}
\arguments{
\item{object}{a RandomForest object}

\item{plot}{what should be plotted? Residuals? model plot? All of them?}

\item{formula}{A flexplot-style formula}

\item{...}{Other arguments passed to flexplot}
}
\description{
Visualize a RandomForest model
}
