% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{youtube_plot}
\alias{youtube_plot}
\title{Create a 1080p plot for YouTube}
\usage{
youtube_plot(p, name = "plot", path = "")
}
\arguments{
\item{p}{the ggplot object}

\item{name}{the name of the file (excluding the ".jpg")}

\item{path}{the path to the folder}
}
\description{
Create a 1080p plot for YouTube
}
