% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/added.plot.R
\name{added.plot}
\alias{added.plot}
\title{Create an added variable plot}
\usage{
added.plot(formula, data, ...)
}
\arguments{
\item{formula}{A formula with exactly two predictors and an outcome variable}

\item{data}{The dataset used}

\item{...}{Other parameters passed to flexplot}
}
\description{
Create an added variable plot
}
\details{
This function first residualizes the outcome variable based on the first variables listed in the formula.
The mean of the outcome variable is then added to the residuals (to maintain the interpretation of the variable),
then the function plots the residuals against the second variable.
}
\examples{
data(exercise_data)
added.plot(weight.loss~motivation + therapy.type, data=d)
}
\seealso{
\code{\link{flexplot}}
}
\author{
Dustin Fife
}
