% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_fits.R
\name{compare.fits}
\alias{compare.fits}
\title{Compare the fits of two models}
\usage{
compare.fits(
  formula,
  data,
  model1,
  model2 = NULL,
  return.preds = F,
  report.se = F,
  re = F,
  pred.type = "response",
  num_points = 50,
  clusters = 3,
  ...
)
}
\arguments{
\item{formula}{A formula that can be used in flexplot. The variables inside must not include variables outside the fitted models.}

\item{data}{The dataset containing the variables in formula}

\item{model1}{The fitted model object (e.g., lm) containing the variables specified in the formula}

\item{model2}{The second fitted model object (e.g., lm) containing the variables specified in the formula}

\item{return.preds}{Should the function return the predictions instead of a graphic? Defaults to F}

\item{report.se}{Should standard errors be reported alongside the estimates? Defaults to F.}

\item{re}{Should random effects be predicted? Only applies to mixed models. Defaults to F.}

\item{pred.type}{What type of predictions should be outputted? This is mostly for \code{glm} models. Defaults to "response."}

\item{num_points}{Number of points used for predictions. Larger numbers = slower algorithm, but smoother predictions.}

\item{clusters}{For visualizing mixed models, this specifies the number of clusters to display}

\item{...}{Other parameters passed to flexplot}
}
\value{
Either a graphic or the predictions for the specified model(s)
}
\description{
This function takes two fitted models as input and plots them to visually compare how the two differ in terms of fit.
It can take a \code{glm}, \code{rlm}, \code{lm}, and \code{randomForest} model (and maybe others as well). The function takes
a \code{\link{flexplot}}-like formula as input.
}
\examples{
data(exercise_data)
mod1 = lm(weight.loss~therapy.type + motivation, data=exercise_data)
mod2 = lm(weight.loss~therapy.type * motivation, data=exercise_data)
compare.fits(weight.loss~therapy.type | motivation, data=exercise_data, mod1, mod2)
}
\author{
Dustin Fife
}
