% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{estimates.default}
\alias{estimates.default}
\title{Output APA style statistical significance from an object}
\usage{
\method{estimates}{default}(object, mc = TRUE)
}
\arguments{
\item{object}{a object}

\item{mc}{Should model comparisons be performed?}
}
\value{
One or more objects containing parameter estimates and effect sizes
}
\description{
Output APA style statistical significance from an object
}
