% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.glm}
\alias{estimates.glm}
\title{Report glm object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{glm}(object, mc = FALSE)
}
\arguments{
\item{object}{a glm object}

\item{mc}{Should model comparisons be performed? Currently not used}
}
\value{
One or more objects containing parameter estimates and effect sizes. 

For logistic regression, it will return the following:
\describe{
  \item{`raw.coefficients`}{The raw log-odds coefficients.}
  \item{`OR`}{Odds ratios (exp(beta)).}
  \item{`inverse.OR`}{Inverse odds ratios (1 / OR).}
  \item{`standardized.OR`}{Standardized odds ratios (exp(beta)).}
  \item{`inverse.standardized.OR`}{Standardized version of the inverse odds ratios (1 / OR).}
  \item{`instantaneous_slope`}{The slope of the probability curve at p = 0.5.}
  \item{`intercept_threshold`}{The value of X where p = 0.5.}
  \item{`Prediction Difference (+/- 1 SD)`}{Change in predicted probability from -1 to +1 SD.}
}
}
\description{
Report glm object Estimates
}
