% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{estimates.RandomForest}
\alias{estimates.RandomForest}
\title{Report RandomForest object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{RandomForest}(object, mc = TRUE)
}
\arguments{
\item{object}{a RandomForest object}

\item{mc}{Should model comparisons be performed? 
Currently not implemented for RandomForest objects}
}
\value{
One or more objects containing parameter estimates and effect sizes
}
\description{
Report RandomForest object Estimates
}
