% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.zeroinfl}
\alias{estimates.zeroinfl}
\title{Report zeroinfl object Estimates (effect sizes and parameters)}
\usage{
\method{estimates}{zeroinfl}(object, mc = FALSE)
}
\arguments{
\item{object}{a zeroinfl object}

\item{mc}{Should model comparisons be performed? Currently not used}
}
\description{
Report zeroinfl object Estimates
}
