% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexplota.h.R
\name{flexplota}
\alias{flexplota}
\title{Flexplot}
\usage{
flexplota(data, out, preds, given, se = TRUE, line = "Loess",
  ghost = TRUE, thirdeye = FALSE, diff = FALSE,
  plmethod = "Jittered-density plot", resid = FALSE, suppr = FALSE,
  center = "Median + quartiles", alpha = 50, sample = 100,
  jittx = 0.2, jitty = 0, bins = 3)
}
\arguments{
\item{data}{.}

\item{out}{.}

\item{preds}{.}

\item{given}{.}

\item{se}{.}

\item{line}{.}

\item{ghost}{.}

\item{thirdeye}{.}

\item{diff}{.}

\item{plmethod}{.}

\item{resid}{.}

\item{suppr}{.}

\item{center}{.}

\item{alpha}{.}

\item{sample}{.}

\item{jittx}{.}

\item{jitty}{.}

\item{bins}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Flexplot
}
