% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexplota.h.R
\name{flexplota}
\alias{flexplota}
\title{Flexplot}
\usage{
flexplota(data, out, preds, se = TRUE, line = "loess",
  center = "median + quartiles")
}
\arguments{
\item{data}{.}

\item{out}{.}

\item{preds}{.}

\item{se}{.}

\item{line}{.}

\item{center}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Flexplot
}
