% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_flexplot.R
\name{modify_labels}
\alias{modify_labels}
\title{Modify the labels in a flexplot (or ggplot) graphic}
\usage{
modify_labels(
  p,
  y = NULL,
  x = NULL,
  color = NULL,
  row_panels = NULL,
  col_panels = NULL
)
}
\arguments{
\item{p}{The flexplot (or ggplot) object}

\item{y}{The new label for the variable on the y axis}

\item{x}{The new label for the variable on the x axis}

\item{row_panels}{The new labels for the variable in row panels}

\item{col_panels}{The new labels for the variable in column panels}

\item{colour}{The new label for the variable in the legend. 
(This will also modify the underlying shape/line arguments)}
}
\value{

}
\description{
Modify the labels in a flexplot (or ggplot) graphic
}
