% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_residual.R
\name{partial_residual_plot}
\alias{partial_residual_plot}
\title{Partial Residual Plot}
\usage{
partial_residual_plot(
  plot_formula,
  lm_formula = NULL,
  model = NULL,
  data = NULL,
  added_term = NULL,
  suppress_model = F,
  add_mean = TRUE,
  ...
)
}
\arguments{
\item{plot_formula}{a flexplot-style formula for plotting}

\item{lm_formula}{Optional. A lm-style formula for the model being fit}

\item{model}{Optional. A lm model}

\item{data}{The dataset}

\item{added_term}{a formula, which specifies which terms should be de-residualized. By default,
it will use the residuals from `model`.}

\item{suppress_model}{A boolean. Should the model be suppressed? Defaults to T.}

\item{...}{Other arguments passed to flexplot}
}
\value{
a partial residual plot
}
\description{
Partial Residual Plot
}
\examples{
partial_residual_plot(weight.loss~therapy.type,
   lm_formula = weight.loss~therapy.type + motivation,
   data=exercise_data)
}
