% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jitterd.R
\name{position_jitterdodged}
\alias{position_jitterdodged}
\title{Simultaneously dodge and jitter}
\usage{
position_jitterdodged(
  jitter.width = 0.2,
  jitter.height = 0,
  dodge.width = 0.75,
  seed = NA
)
}
\arguments{
\item{jitter.width}{degree of jitter in x direction. Defaults to 40\% of the
resolution of the data.}

\item{jitter.height}{degree of jitter in y direction. Defaults to 0.}

\item{dodge.width}{The degree of dodging}

\item{seed}{Random seed.}
}
\description{
This is primarily used for aligning points generated through
`geom_point()` with dodged boxplots (e.g., a `geom_boxplot()` with
a fill aesthetic supplied).
}
\seealso{
Other position adjustments: 
\code{\link{position_jitterd}()}
}
\concept{position adjustments}
