% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{print.estimates}
\alias{print.estimates}
\title{Print estimates Summary}
\usage{
\method{print}{estimates}(x, ...)
}
\arguments{
\item{x}{an estimates object}

\item{...}{ignored}
}
\value{
A printed list of estimates
}
\description{
Print estimates Summary
}
