% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{print.lmer_estimates}
\alias{print.lmer_estimates}
\title{Print lmer_estimates Summary}
\usage{
\method{print}{lmer_estimates}(x, ...)
}
\arguments{
\item{x}{an lmer_estimates object}

\item{...}{ignored}
}
\value{
A printed list of estimates
}
\description{
Print a lmer_estimates object
}
