% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.lmerMod}
\alias{visualize.lmerMod}
\title{Visualize a fitted lmerMod model}
\usage{
\method{visualize}{lmerMod}(object, plot = c("residuals", "all",
  "model"), formula = NULL, sample = 3, ...)
}
\arguments{
\item{object}{a lmer object}

\item{plot}{what should be plotted? Residuals? Model plot? All of them?}

\item{formula}{A flexplot-style formula}

\item{sample}{The number of individuals' datapoints to sample as raw data. Defaults to 3}

\item{...}{Other arguments passed to flexplot}
}
\description{
Visualize a fitted lmerMod model
}
