% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.lmerMod}
\alias{visualize.lmerMod}
\title{Visualize a fitted model}
\usage{
\method{visualize}{lmerMod}(object, plot = c("residuals", "all",
  "bivariate"), formula = NULL, ...)
}
\arguments{
\item{object}{a lmer object}

\item{plot}{what should be plotted? Residuals? Bivariate plot? All of them?}

\item{formula}{A flexplot-style formula}

\item{...}{Other arguments passed to flexplot}
}
\description{
Visualize a fitted model
}
