% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_lmer.R
\name{visualize.lmerMod}
\alias{visualize.lmerMod}
\title{Visualize a fitted lmerMod model}
\usage{
\method{visualize}{lmerMod}(
  object,
  plot = c("all", "residuals", "model"),
  formula = NULL,
  sample = 3,
  plots.as.list = FALSE,
  ...
)
}
\arguments{
\item{object}{a lmer object}

\item{plot}{what should be plotted? Residuals? Model plot? All of them?}

\item{formula}{A flexplot-style formula}

\item{sample}{The number of individuals' datapoints to sample as raw data. Defaults to 3}

\item{plots.as.list}{Should the plots be returned as a list? Defaults to FALSE.}

\item{...}{Other arguments passed to flexplot}
}
\value{
a plot containing a visual of the chosen model
}
\description{
Visualize a fitted lmerMod model
}
