% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize a fitted model}
\usage{
visualize(object, plot = c("all", "residuals", "model"), formula = NULL, ...)
}
\arguments{
\item{object}{a fitted object}

\item{plot}{what should be plotted? Residuals? Bivariate plot? All of them?}

\item{formula}{A flexplot-style formula}

\item{...}{Other arguments passed to flexplot}
}
\value{
a plot containing a visual of the chosen model
}
\description{
Visualize a fitted model
}
